refactor: UI/이벤트 파일 분리
