+++ src/main.js
@@ -1,34 +1,9 @@
-import { ErrorPage } from "./components/pages/ErrorPage";
-import { LoginPage } from "./components/pages/LoginPage";
-import { MainPage } from "./components/pages/MainPage";
-import { ProfilePage } from "./components/pages/ProfilePage";
-
-const routes = {
-  "/": MainPage,
-  "/login": LoginPage,
-  "/profile": ProfilePage,
-};
-
-const router = () => {
-  const path = window.location.pathname;
-  const render = routes[path] || ErrorPage;
-  document.body.innerHTML = render();
-};
-
-const handleNavigation = (e) => {
-  const selected = e.target.closest("a");
-  if (selected && selected.getAttribute("href")) {
-    e.preventDefault();
-    const path = selected.getAttribute("href");
-    history.pushState({}, "", path);
-    router();
-  }
-};
+import { router } from "./router/router.js";
+import { initNavigation } from "./events/navigation.js";
 
 document.addEventListener("DOMContentLoaded", () => {
   router();
-  document.body.addEventListener("click", handleNavigation);
+  initNavigation();
 });
 
-// 뒤로가기/앞으로가기
 window.addEventListener("popstate", router);
