+++ src/router/router.js
@@ -1,0 +1,16 @@
+import { ErrorPage } from "../components/pages/ErrorPage";
+import { LoginPage } from "../components/pages/LoginPage";
+import { MainPage } from "../components/pages/MainPage";
+import { ProfilePage } from "../components/pages/ProfilePage";
+
+const routes = {
+  "/": MainPage,
+  "/login": LoginPage,
+  "/profile": ProfilePage,
+};
+
+export const router = () => {
+  const path = window.location.pathname;
+  const render = routes[path] || ErrorPage;
+  document.body.innerHTML = render();
+};
